IOS-7170: Move hide token alert logic to builder
